fix: 기본 page는 props 커스텀 못함
